requirements.txt
asgiref==3.4.1
cx-Oracle==8.3.0
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.6
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
